Expanded survival documentation and protocols. Added Natural Disasters, Water/Food procurement, Shelter types, and Primitive Navigation. Fixed .gitignore for assets.

+++ survival_aid_app/lib/services/model_setup_service.dart
@@ -112,7 +112,7 @@ class ModelSetupService extends Notifier<ModelSetupState> {
         token: huggingFaceToken,
       ).withProgress((progress) {
         // progress is an int representing percentage (0-100)
-        final p = (progress as int);
+        final p = progress;
         state = state.copyWith(
           downloadProgress: p / 100.0,
           statusMessage: 'Downloading Gemma ($p%)...',
